requirements.txt
crc16
ipaddr
django-excel-response
xlwt